Moved haversine testing from perturbation_auf_wrapper to misc_functions, introducing a new 'wrapper' fortran subroutine to access the otherwise uncallable haversine formula.
